% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSstack_predict.R
\name{BSstack_predict}
\alias{BSstack_predict}
\title{Predict using a set of Stacked Random Forest models.}
\usage{
BSstack_predict(BSmodel, Xi)
}
\arguments{
\item{BSmodel}{List containing the individual Random Forest models, their weights, and feature names. Generated using BSstack function.}

\item{Xi}{NxM datatable containing input features to be predicted. Must contain all features used in the individual RF models.}
}
\value{
NxL+1 matrix where L is the number of individual RF models. Predictions for the ith RF model is found in the ith column of this matrix while predictions for the stacked model is in the final column.
}
\description{
Gives predictions for a linear bootstrapped stacked Random Forest predictors. Gives the predictions
of each individual model as well as the linearlly combined predictions.
}
\details{
Required Packages:
  randomForest
}
